# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/users-regions": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store a newly created Customer in storage",
                "description": "Store User",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "region",
                        "in": "body",
                        "description": "Add user region.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/UserRegions"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/User"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/update-push-notification": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update the specified User's push notification in storage",
                "description": "Update User's push notification",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "push_notification",
                        "in": "query",
                        "description": "Set Push Notification: 0,1",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/User"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "UserRegions": {
            "required": [
                "region_id"
            ],
            "properties": {
                "region_id": {
                    "description": "Region id",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "region_reminder": {
                    "description": "region reminder",
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "User": {
            "required": [
                "name",
                "email",
                "password"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
    }
}